update demo file name
