+++ docs/demo/basic.md
@@ -1,3 +1,3 @@
-## stupid
+## simple
 
 <code src="../examples/simple.jsx">
