+++ docs/demo/controlled.md
@@ -1,0 +1,3 @@
+## default
+
+<code src="../examples/controlled.jsx">
