--- docs/demo/default.md
@@ -1,3 +1,0 @@
-## default
-
-<code src="../examples/default.jsx">
