--- docs/demo/stupid.md
@@ -1,3 +1,0 @@
-## stupid
-
-<code src="../examples/stupid.jsx">
