--- docs/examples/stupid.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import Pagination from 'rc-pagination';
-import '../../assets/index.less';
-
-export default () => <Pagination simple defaultCurrent={1} total={50} />;
